Rework item code, and start delegating LookupForm form building to some interface.
