+++ .idea/compiler.xml
@@ -35,7 +35,6 @@
       <module name="domui-hello" target="1.5" />
       <module name="domui-hibernate-generator" target="1.8" />
       <module name="domui-utilities" target="1.5" />
-      <module name="property-annotations-processor" target="1.8" />
       <module name="to.etc.alg" target="1.8" />
       <module name="to.etc.db" target="1.8" />
       <module name="to.etc.dbcompare" target="1.8" />
