+++ to.etc.domui/src/main/java/to/etc/domui/component/lookupform2/lookupcontrols/LookupControlRegistry2.java
@@ -90,4 +90,13 @@ public class LookupControlRegistry2 {
 		}
 		return best;
 	}
+
+	@Nullable
+	public FactoryPair<?> findControlPair(SearchPropertyMetaModel spm) {
+		ILookupFactory<?> factory = findFactory(spm);
+		if(null == factory)
+			return null;
+		FactoryPair<?> controlPair = factory.createControl(spm);
+		return controlPair;
+	}
 }
